requirements.txt
attrs
beautifulsoup4
lxml
pandas
retrying
selenium
statsmodels
traceback
